fix(webpack): browser and server have specific endpoints
They reach in for toCollectionName differently
