+++ package.json
@@ -14,7 +14,7 @@
     "type": "MIT",
     "url": "https://www.opensource.org/licenses/mit-license.php"
   },
-  "main": "lib",
+  "main": "lib/index.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/saiichihashimoto/mongoose-normalizr.git"
@@ -34,6 +34,7 @@
       "helpMessage": "Install commitizen and use `git cz` instead of `git commit` https://github.com/commitizen/cz-cli#installing-the-command-line-tool"
     }
   },
+  "browser": "lib/index.web.js",
   "dependencies": {
     "mongoose": "^4.13.7",
     "normalizr": "^3.2.4"
