+++ src/index.web.js
@@ -1,0 +1,4 @@
+import mongoose from 'mongoose';
+import findNormalizrs from './findNormalizrs';
+
+export default findNormalizrs(mongoose.utils.toCollectionName);
